requirements.txt
pyautogen
huggingface_hub
python-dotenv
langchain
openai
pypdf
python-docx

